+++ src/grandorgue/midi/elements/GOMidiSender.cpp
@@ -222,19 +222,19 @@ void GOMidiSender::ToYaml(YAML::Node &yamlNode, GOMidiMap &map) const {
     eventNode[WX_EVENT_TYPE] = MIDI_SEND_TYPES.GetName(e.type);
 
     if (hasChannel(e.type))
-      eventNode[WX_CHANNEL] = e.channel;
+      eventNode[WX_CHANNEL] = (int)e.channel;
     if (HasKey(e.type))
-      eventNode[WX_KEY] = e.key;
+      eventNode[WX_KEY] = (int)e.key;
     if (isNote(e.type))
-      eventNode[WX_USE_NOTE_OFF] = e.useNoteOff;
+      eventNode[WX_USE_NOTE_OFF] = (int)e.useNoteOff;
     if (hasLowValue(e.type))
-      eventNode[WX_LOW_VALUE] = e.low_value;
+      eventNode[WX_LOW_VALUE] = (int)e.low_value;
     if (hasHighValue(e.type))
-      eventNode[WX_HIGH_VALUE] = e.high_value;
+      eventNode[WX_HIGH_VALUE] = (int)e.high_value;
     if (hasStart(e.type))
-      eventNode[WX_START] = e.start;
+      eventNode[WX_START] = (int)e.start;
     if (hasLength(e.type))
-      eventNode[WX_LENGTH] = e.length;
+      eventNode[WX_LENGTH] = (int)e.length;
     yamlNode.push_back(eventNode);
   }
 }
